requirements.txt
# Core dependencies
numpy>=1.24.0
asyncio-mqtt>=0.13.0

# Audio processing
pyaudio>=0.2.11
soundfile>=0.12.1
librosa>=0.10.0

# Speech-to-Text
faster-whisper>=0.9.0
torch>=2.0.0
torchaudio>=2.0.0

# Character transformation (LLM)
llama-cpp-python>=0.2.0
transformers>=4.30.0

# Text-to-Speech
coqui-tts>=0.15.0
espeak-ng>=1.51

# GPU acceleration
nvidia-ml-py>=12.0.0

# Configuration and data handling
pydantic>=2.0.0
pyyaml>=6.0
jsonschema>=4.17.0

# UI framework
tkinter-tooltip>=2.0.0
# Alternative: PyQt6>=6.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and monitoring
loguru>=0.7.0
psutil>=5.9.0